attribution/requirements.txt
lightgbm>=4.1.0
shap>=0.44.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
joblib>=1.3.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
